requirements.txt
kafka-python >= 2.1.3
keras~=3.9.2
typing_extensions~=4.13.2
numpy~=2.2.5